Apply suggestions from code review

+++ src/common/mmo.hpp
@@ -724,12 +724,12 @@ struct guild_castle {
 	int mapindex;
 	char castle_name[NAME_LENGTH];
 	char castle_event[NPC_NAME_LENGTH];
-	int siege_type;
+	uint8 siege_type;
 	bool enable_client_warp;
-	int warp_x;
-	int warp_y;
-	int zeny;
-	int zeny_siege;
+	uint16 warp_x;
+	uint16 warp_y;
+	uint32 zeny;
+	uint32 zeny_siege;
 	int guild_id;
 	int economy;
 	int defense;
